+++ MDBilibili/app/src/main/java/me/qixingchen/mdbilibili/ui/activity/PlayerActivity.java
@@ -27,12 +27,12 @@ import master.flame.danmaku.danmaku.parser.IDataSource;
 import master.flame.danmaku.danmaku.parser.android.BiliDanmukuParser;
 import me.qixingchen.mdbilibili.R;
 import me.qixingchen.mdbilibili.app.BilibiliApplication;
-import me.qixingchen.mdbilibili.logger.Log;
 import me.qixingchen.mdbilibili.model.VideoM;
 import me.qixingchen.mdbilibili.network.DownloadXML;
 import me.qixingchen.mdbilibili.network.GetXMLinfo;
 import me.qixingchen.mdbilibili.network.RetrofitNetworkAbs;
 import me.qixingchen.mdbilibili.ui.base.BaseActivity;
+import me.qixingchen.mdbilibili.utils.Log;
 import me.qixingchen.mdbilibili.widget.MediaController;
 import me.qixingchen.mdbilibili.widget.VideoView;
 import tv.danmaku.ijk.media.player.IMediaPlayer;
